refactored some of the obstacles into a parent class Obstacle
