+++ src/edu/ucsb/cs56/projects/games/country_runner/Sheep.java
@@ -7,7 +7,7 @@ import java.lang.Math;
  * @version cs56, W14, proj2
  *
  */
-public class Sheep extends Sprite
+public class Sheep extends Obstacle
 {
     //initialXPosition goes into the super constructor
     private static final double initialXPosition = -100.0;
@@ -24,10 +24,6 @@ public class Sheep extends Sprite
     //The sequence that holds the running images
     private SpriteSequence runningSequence;
 
-    //holds score for amount of times user jumped over this object
-    private int score;
-
-
     /** Default Constructor makes the Sheep.
      * sets up the spriteSheet and fills the
      * sequences with images from it
@@ -36,14 +32,9 @@ public class Sheep extends Sprite
     {
     	//Call super constructor
     	super(100, 50, initialXPosition, "sheepSheet");
-
-	//initialize
-        score = 0;
-	setDifficulty(difficulty);
-
+		setDifficulty(difficulty);
 		//Initilize the sequence
 		runningSequence = new SpriteSequence();
-
 		//Fill the sequence
 		//NOTE: we have to explicitly say the number of
 		//images in the sequence
@@ -52,9 +43,6 @@ public class Sheep extends Sprite
 		{
 			this.runningSequence.addImage(getSubImage(i, 0));
 		}
-
-        score = 0;
-
     }
 
 	/** setDifficulty
@@ -93,25 +81,27 @@ public class Sheep extends Sprite
     	//Right now, using the actual size of the window.
     	//Will want to change this later...
 	//todo: factor timer function out; make it more time precise 
-	if (occurance > 0) {
-	    if (waiting) {
-		if (counter > 0)
-		    counter--;
-		else
-		    waiting = false;
-	    }
-	    else
+		if (occurance > 0)
 		{
-		    this.setX(this.getX() + speed);
-		    if (this.getX() == 600)
+		    if (waiting)
+		    {
+				if (counter > 0)
+				    counter--;
+				else
+				    waiting = false;
+			}
+		    else
 			{
-			    this.setX(initialXPosition);
-			    score++;
-			    counter = randomWithRange(occurance, occurance+100);
-			    waiting = true;
+			    this.setX(this.getX() + speed);
+			    if (this.getX() == 600)
+				{
+				    this.setX(initialXPosition);
+				    score++;
+				    counter = randomWithRange(occurance, occurance+100);
+				    waiting = true;
+				}
 			}
 		}
-	}
     }
 
 	/** updateCurrentImage
