+++ src/edu/ucsb/cs56/projects/games/country_runner/Snail.java
@@ -7,10 +7,8 @@ import java.lang.Math;
  * @version cs56, W14, proj2
  *
  */
-public class Snail extends Sprite
+public class Snail extends Obstacle
 {
-    //initialXPosition goes into the super constructor
-    private static final double initialXPosition = -100.0;
 
     //speed and occruance depends on difficulty level
     private static double speed = 10.0;
@@ -20,13 +18,11 @@ public class Snail extends Sprite
     private static int occurance = 12;
     private static int counter;
     private static Boolean waiting;
-    
+    private static final double initialXPosition = -100.0;
     //The sequence that holds the running images
     private SpriteSequence runningSequence;
 
     //holds score for amount of times user jumped over this object
-    private int score;
-
 
     /** Default Constructor makes the Sheep.
      * sets up the spriteSheet and fills the
@@ -36,10 +32,7 @@ public class Snail extends Sprite
     {
     	//Call super constructor
     	super(100, 50, initialXPosition, "snailSheet");
-
-	//initialize
-        score = 0;
-	setDifficulty(difficulty);
+		setDifficulty(difficulty);
 
 		//Initilize the sequence
 		runningSequence = new SpriteSequence();
@@ -52,9 +45,6 @@ public class Snail extends Sprite
 		{
 			this.runningSequence.addImage(getSubImage(i, 0));
 		}
-
-        score = 0;
-
     }
 
 	/** setDifficulty
@@ -92,46 +82,27 @@ public class Snail extends Sprite
     {
     	//Right now, using the actual size of the window.
     	//Will want to change this later...
-	//todo: factor timer function out; make it more time precise 
-	if (occurance > 0) {
-	    if (waiting) {
-		if (counter > 0)
-		    counter--;
-		else
-		    waiting = false;
-	    }
-	    else
-		{
-		    this.setX(this.getX() + speed);
-		    if (this.getX() == 600)
+		//todo: factor timer function out; make it more time precise 
+		if (occurance > 0) {
+		    if (waiting) {
+				if (counter > 0)
+				    counter--;
+				else
+				    waiting = false;
+			    }
+		    else
 			{
-			    this.setX(initialXPosition);
-			    score++;
-			    counter = randomWithRange(occurance, occurance+100);
-			    waiting = true;
+			    this.setX(this.getX() + speed);
+			    if (this.getX() == 600)
+				{
+				    this.setX(initialXPosition);
+				    score++;
+				    counter = randomWithRange(occurance, occurance+100);
+				    waiting = true;
+				}
 			}
 		}
-	}
     }
-
-	/** updateCurrentImage
-	 * Moves to the next image in the running sequence
-     */
-	public void updateCurrentImage()
-	{
-		setCurrentImage(runningSequence.getNextImage());
-	}
-
-    public int getScore() {
-        return score;
-    }
-
-    int randomWithRange(int min, int max)
-    {
-	int range = (max - min) + 1;     
-	return (int)(Math.random() * range) + min;
-    }
-    
 }
